Refactored raw upload logic by decoupling file processing and logging into `ingest_utils.py`. Removed unused `raw_upload.py`. Updated imports and references accordingly.
